fix test_diffraction_vectors2d
